python/google-adk/sample-agent/pyproject.toml
[project]
name = "sample-google-adk"
version = "0.1.0"
description = "Sample Google ADK Agent using Microsoft Agent 365 SDK"
authors = [
    { name = "Microsoft", email = "support@microsoft.com" }
]
dependencies = [
    # Google ADK -- official package
    "google-adk",

    # Microsoft Agents SDK - Official packages for hosting and integration
    "microsoft-agents-hosting-aiohttp",
    "microsoft-agents-hosting-core",
    "microsoft-agents-authentication-msal",
    "microsoft-agents-activity",

    # Core dependencies
    "python-dotenv",
    "aiohttp",

    # HTTP server support for MCP servers
    "uvicorn[standard]>=0.20.0",
    "fastapi>=0.100.0",

    # HTTP client
    "httpx>=0.24.0",

    # Data validation
    "pydantic>=2.0.0",

    # Additional utilities
    "typing-extensions>=4.0.0",

    # Microsoft Agent 365 SDK packages
    "microsoft_agents_a365_tooling >= 0.1.0",
]
requires-python = ">=3.11"

# Package index configuration
# PyPI is the default/primary source, local packages are fallback
[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[project.optional-dependencies]
dev = [
    # For development and testing
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# Don't include any Python modules in the package since this is a sample/script collection
py-modules = []

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["build*", "dist*", "venv*"]
